@@ -2,11 +2,11 @@
 name: musily
 description: "A great music app."
 publish_to: "none" # Set to 'none' to prevent publishing to pub.dev
-version: 4.0.1
+version: 4.0.3
 
 # Flutter SDK Environment
 environment:
-  sdk: ">=3.4.1 <4.0.0"
+  sdk: ">=3.7.0 <4.0.0"
 
 # Dependencies
 dependencies:
@@ -23,15 +23,14 @@ dependencies:
   dio: ^5.5.0+1
 
   # YouTube APIs
-  youtube_explode_dart: 
-    git:
-      url: https://github.com/Hexer10/youtube_explode_dart.git
-      ref: e519db65ad0b0a40b12f69285932f9db509da3cf
-  dart_ytmusic_api: ^1.3.2
+  youtube_explode_dart:
+    git: https://github.com/Coronon/youtube_explode_dart
+  dart_ytmusic_api:
+    git: https://github.com/MusilyApp/dart_ytmusic_api
 
   # UI and Animations
   loading_animation_widget: ^1.2.1
-  skeletonizer: ^1.4.1+1
+  skeletonizer: ^2.1.0+1
   dotted_border: ^2.1.0
   flutter_svg: ^2.0.10+1
   percent_indicator: ^4.2.3
@@ -62,7 +61,7 @@ dependencies:
   # Localization
   flutter_localizations:
     sdk: flutter
-  intl: ^0.19.0
+  intl: ^0.20.2
 
   # Permissions and App Links
   permission_handler: ^11.3.1
@@ -92,8 +91,9 @@ dependencies:
   process_run: ^0.14.2
   dart_ipc: ^1.0.0
   package_info_plus: ^8.3.0
+  gpt_markdown: ^1.1.4
+  flutter_curl: ^0.7.0
   lucide_icons_flutter: ^3.1.4
-  gpt_markdown: 1.0.0
 
 # Development Dependencies
 dev_dependencies:
