@@ -5,8 +5,8 @@ import 'package:musily/core/presenter/ui/utils/ly_navigator.dart';
 import 'package:musily/core/presenter/widgets/app_builder.dart';
 import 'package:musily/core/utils/display_helper.dart';
 import 'package:musily/features/settings/presenter/controllers/settings/settings_controller.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:musily/features/settings/presenter/controllers/settings/settings_data.dart';
+import 'package:musily/l10n/app_localizations.dart';
 
 extension MusilyBuildContext on BuildContext {
   ThemeMode get themeMode {
@@ -26,9 +26,11 @@ extension MusilyBuildContext on BuildContext {
       BuildContext context,
       BaseControllerEvent event,
       SettingsData data,
-    )? eventListener,
+    )?
+    eventListener,
     bool allowAlertDialog,
-  }) get settingsBuilder => Modular.get<SettingsController>().builder;
+  })
+  get settingsBuilder => Modular.get<SettingsController>().builder;
 
   AppLocalizations get localization {
     return AppLocalizations.of(this)!;
@@ -39,16 +41,17 @@ extension MusilyBuildContext on BuildContext {
   BuildContext get showingPageContext {
     final contextManager = ContextManager();
 
-    final foundContext = contextManager.contextStack
-        .where(
-          (e) => [
-            'SectionsPage',
-            'SearchPage',
-            'DownloaderPage',
-            'LibraryPage',
-          ].contains(e.key),
-        )
-        .firstOrNull;
+    final foundContext =
+        contextManager.contextStack
+            .where(
+              (e) => [
+                'SectionsPage',
+                'SearchPage',
+                'DownloaderPage',
+                'LibraryPage',
+              ].contains(e.key),
+            )
+            .firstOrNull;
 
     return foundContext?.context ?? this;
   }
