@@ -3,7 +3,7 @@ import 'dart:ui';
 import 'package:audio_service/audio_service.dart';
 import 'package:just_audio_media_kit/just_audio_media_kit.dart';
 import 'package:musily/features/downloader/presenter/controllers/downloader/downloader_controller.dart';
-import 'package:musily/features/player/data/services/musily_audio_handler.dart';
+import 'package:musily/features/player/domain/entities/musily_audio_handler.dart';
 import 'package:musily/features/player/data/services/musily_audio_handler_impl.dart';
 import 'package:musily/features/player/data/services/musily_player.dart';
 
